Serverless, Edge, Fog/template_http/code/requirements.txt
google-cloud-storage==1.33.0
flask==1.1.2